+++ source/mysql/protocol/packets.d
@@ -13,6 +13,7 @@ import std.string;
 import std.traits;
 import std.variant;
 
+import mysql.commands : ColumnSpecialization, CSN;
 import mysql.exceptions;
 import mysql.protocol.constants;
 import mysql.protocol.extra_types;